fix. 스타일북 > 스타일북 Model 값 패키지 수정
